# Conflicts:
#	backend/src/main/java/com/synerge/order101/purchase/model/entity/Purchase.java
#	backend/src/main/java/com/synerge/order101/purchase/model/service/PurchaseServiceImpl.java
#	backend/src/main/java/com/synerge/order101/warehouse/model/repository/WarehouseInventoryRepository.java

+++ backend/src/main/java/com/synerge/order101/order/model/dto/StoreOrderCreateRequest.java
@@ -35,6 +35,6 @@ public class StoreOrderCreateRequest {
         private Long productId;
 
         @NonNull
-        private BigDecimal orderQty;
+        private Integer orderQty;
     }
 }
